add: Visual toolbar and reorganized files
